requirements.txt
# My Blue App - 依赖列表

# 蓝牙通信核心库
pybluez>=0.23
bleak>=0.21.0

# GUI 框架
PyQt5>=5.15.0

# 异步IO支持
# asyncio (Python 3.7+ 内置)

# 配置文件解析
pyyaml>=6.0

# 日志
# logging (Python 标准库)

# 开发和测试工具
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# 代码质量工具
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.4.0

# 类型存根
types-PyYAML>=6.0.12

# 文档生成
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0